pyproject.toml
[project]
name = "pdf-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "fastmcp>=2.8.1",
    "pdfminer-six>=20250506",
]
